opt: Move logical and physical property files to new package
Create new props package that consolidates the definitions of both
logical and physical properties. Properties are heavily used across
the entire optimizer codebase, and so centralizing them in their own
package makes sense.

Note that the package just contains the definitions of the properties,
not the code that populates them. That code will be contained in multiple
other packages, depending on how the property is meant to be used.

Release note: None
